+++ gtk/gtkmountoperation.c
@@ -1088,7 +1088,7 @@ add_pid_to_process_list_store (GtkMountOperation              *mount_operation,
                                     &pixbuf);
 
   if (name == NULL)
-    name = g_strdup_printf (_("Unknown Application (PID %d)"), pid);
+    name = g_strdup_printf (_("Unknown Application (PID %d)"), (int) pid);
 
   if (command_line == NULL)
     command_line = g_strdup ("");
